removed warnings when compiling for 64 bit
This is actually just an intermediate solution. It would probably be better to change all Sparrow APIs to use 'NSInteger' and 'NSUInteger' instead of 'int' and 'uint'. But for now, this solution works.

+++ samples/demo/src/Classes/BenchmarkScene.m
@@ -119,7 +119,7 @@
     _started = NO;
     _startButton.visible = YES;
     
-    int frameRate = Sparrow.currentController.framesPerSecond;
+    int frameRate = (int)Sparrow.currentController.framesPerSecond;
     
     NSLog(@"benchmark complete!");
     NSLog(@"fps: %d", frameRate);
