+++ sparrow/src/Classes/SPSprite.m
@@ -48,7 +48,7 @@
     if (_flattenedContents)
     {
         [support finishQuadBatch];
-        [support addDrawCalls:_flattenedContents.count];
+        [support addDrawCalls:(int)_flattenedContents.count];
         
         SPMatrix *mvpMatrix = support.mvpMatrix;
         float alpha = support.alpha;
